Fix file and archive functionality
According to the Oozie schema the file and archive
nodes are inside the action operation node, i.e.:

<action>
  <pig>
    <file>
    <archive>

However we expected them to be one level up:

<action>
  <pig>
  </pig>
  <file>
  <archive>

This commit fixed that and amends the Pig example
workflow to follow the correct schema.

---

Additionally, there was an error in FileMixin where was invoking
add_archive() instead of add_file(). This is fixed.

Added a unit test for parser validating the parsing of file and
archive nodes (on the basis of Pig).