Add XML escaping of dataproc_hadoop_properties in the MapReduce template
During the development of Node notifications we found out that these
properties are not XML-escaped when passed to Dataproc (and they are,
in fact, stored in XML in the Dataproc instance).

This results in errors, such as when one of the parameters is a URL
with "&" signs (which should be escaped to "&amp;" in XML).
Also there's potential to inject XML (e.g. closing tags </value>).

This is likely a bug in Dataproc, which we'll report separately -
however for now we have to escape these properties ourselves.